Purge GORM stuff

+++ api/models/administrator.go
@@ -5,7 +5,7 @@ import (
 )
 
 type Administrator struct {
-	ID        int64     `gorm:"column:id;primaryKey" json:"id"`
+	ID        int32     `gorm:"column:id;primaryKey" json:"id"`
 	CreatedAt time.Time `gorm:"column:created_at;not null" json:"created_at"`
 	UpdatedAt time.Time `gorm:"column:updated_at;not null" json:"updated_at"`
 }
